requirements.txt
tqdm
numpy
opencv-python
pytorch
torchvision
nvdiffrast